requirements.txt
schedule==0.2.1
django-uuidfield==0.5.0
requests>=2.2.1